Moved collision velocity contraint outside the pyopensot.cpp file in order to make the bindings compile when the flag OPENSOT_COMPILE_COLLISION is False. Collisions bindngs are now in pyopensot_collision python lib.
