+++ bindings/python/CMakeLists.txt
@@ -14,20 +14,23 @@ if(${pybind11_FOUND})
 
     message(STATUS "compiling python bindings")
 
+    set(OPENSOT_BINDINGS_SOURCES    pyopensot.cpp
+                                    variables.hpp
+                                    autostack.hpp
+                                    sub.hpp
+                                    aggregated.hpp
+                                    generic.hpp
+                                    basic.hpp
+                                    solver.hpp
+                                    tasks/velocity.hpp
+                                    tasks/acceleration.hpp
+                                    constraints/velocity.hpp
+                                    constraints/acceleration.hpp
+                                    constraints/force.hpp)
+
     include_directories(${CMAKE_CURRENT_SOURCE_DIR})
-    pybind11_add_module(pyopensot pyopensot.cpp
-                                  variables.hpp
-                                  autostack.hpp
-                                  sub.hpp
-                                  aggregated.hpp
-                                  generic.hpp
-                                  basic.hpp
-                                  solver.hpp
-                                  tasks/velocity.hpp
-                                  tasks/acceleration.hpp
-                                  constraints/velocity.hpp
-                                  constraints/acceleration.hpp
-                                  constraints/force.hpp)
+    pybind11_add_module(pyopensot ${OPENSOT_BINDINGS_SOURCES})
+
     target_link_libraries(pyopensot PUBLIC OpenSoT)
 
 
@@ -38,6 +41,25 @@ if(${pybind11_FOUND})
         LIBRARY DESTINATION "~/.local/lib/python${Python3_VERSION_MAJOR}.${Python3_VERSION_MINOR}/site-packages"
         ARCHIVE DESTINATION "lib"
         RUNTIME DESTINATION "bin")
+
+    if(${OPENSOT_COMPILE_COLLISION})
+        set(OPENSOT_BINDINGS_COLLISION_SOURCES pyopensot_collision.cpp constraints/velocity_collision.hpp)
+
+        include_directories(${CMAKE_CURRENT_SOURCE_DIR})
+        pybind11_add_module(pyopensot_collision ${OPENSOT_BINDINGS_COLLISION_SOURCES})
+
+        target_link_libraries(pyopensot_collision PUBLIC OpenSoT)
+
+
+        find_package(Python3 COMPONENTS Interpreter Development REQUIRED)
+
+        install(TARGETS pyopensot_collision
+            COMPONENT python
+            LIBRARY DESTINATION "~/.local/lib/python${Python3_VERSION_MAJOR}.${Python3_VERSION_MINOR}/site-packages"
+            ARCHIVE DESTINATION "lib"
+            RUNTIME DESTINATION "bin")
+    endif()
+
 
 
 else()
